add adminDao adminService adminController admin API endpoints

+++ src/main/java/com/adrian/dao/EmployeeDao.java
@@ -8,8 +8,8 @@ public interface EmployeeDao {
 	
 	public void addEmployee(Employee e);
 	public List<Employee> getAllEmployees();
-	public Employee getEmployeeByID(long id);
-	public void deleteEmployee(long id);
+	public Employee getEmployeeByID(int id);
+	public void deleteEmployee(int id);
 	public void updateEmployee(Employee e);
